move switch, log
